Correct if url, box6Total as Double

+++ app/uk/gov/hmrc/organisationsdetailsapi/domain/integrationframework/VatReturnDetails.scala
@@ -23,7 +23,7 @@ case class IfVatReturn(
                         calendarMonth: Option[Int],
                         liabilityMonth: Option[Int],
                         numMonthsAssessed: Option[Int],
-                        box6Total: Option[Int],
+                        box6Total: Option[Double],
                         returnType: Option[String],
                         source: Option[String]
                       )
